data_pipeline/requirements.txt
# Core dependencies
numpy==1.24.2
pandas==2.0.0
scipy==1.10.1
matplotlib==3.7.1
networkx==3.1
statsmodels==0.14.0
ruptures==1.1.7

# Vector database
qdrant-client==1.4.0

# OpenAI
openai==1.3.0
httpx<0.25.0

# Redis for caching
redis>=4.5.5
# aioredis is now included in redis-py 4.2.0+

# MongoDB
motor>=3.1.1
pymongo>=4.3.3

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
scipy>=1.10.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0

# NLP and text analysis
nltk>=3.8.1
wordcloud>=1.8.2

# Background tasks
celery>=5.2.7

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0 

seaborn==0.12.2
plotly==5.14.1 